+++ src/test/java/edu/hm/hafner/analysis/parser/ErrorProneParserTest.java
@@ -1,5 +1,7 @@
 package edu.hm.hafner.analysis.parser;
 
+import org.junit.jupiter.api.Test;
+
 import edu.hm.hafner.analysis.AbstractParserTest;
 import edu.hm.hafner.analysis.IssueParser;
 import edu.hm.hafner.analysis.Report;
@@ -39,6 +41,20 @@ class ErrorProneParserTest extends AbstractParserTest {
                         + "for (String line : Splitter.onPattern(&quot;\\\\r?\\\\n&quot;).split(stackTrace)) {"
                         + "</code></pre>"
                         + "<p><a href=\"https://errorprone.info/bugpattern/StringSplitter\">See ErrorProne documentation.</a></p>");
+    }
+
+    @Test
+    void shouldFindAllWarnings() {
+        Report report = parse("error-prone-maven.log");
+
+        assertThat(report).hasSize(1);
+        assertThat(report.get(0))
+                .hasFileName("/Users/hafner/Development/jenkins/workspace/Model - Freestyle - New/src/main/java/edu/hm/hafner/analysis/parser/RobocopyParser.java")
+                .hasLineStart(29)
+                .hasColumnStart(45)
+                .hasType("StringSplitter")
+                .hasMessage("String.split(String) has surprising behavior")
+                .hasDescription("Did you mean: <pre><code>String file = matcher.group(4).split(&quot;\\\\s{11}&quot;, -1)[0];</code></pre><p><a href=\"http://errorprone.info/bugpattern/StringSplitter\">See ErrorProne documentation.</a></p>");
     }
 
     @Override